Core 空安全處理

+++ core/lib/src/util/double_util.dart
@@ -2,7 +2,7 @@ import 'dart:math';
 
 class DoubleUtil {
   static double getDecimalsValue(double val, int decimals) {
-    int fac = pow(10, decimals);
+    int fac = pow(10, decimals) as int;
     return (val * fac).round() / fac;
   }
 }
